requirements.txt
# ==============================================================================
# AI Productivity Agent v3.0 — Dependencies
# Source of truth: Design Document v3.0, Section 11 (Migration Guide)
# ==============================================================================

# Core API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# LangChain / LangGraph / Gemini
langchain>=0.2.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0

# Database
sqlmodel>=0.0.18
aiosqlite>=0.20.0
alembic>=1.13.0

# Tools
dateparser>=1.2.0
python-Levenshtein>=0.23.0

# Observability
structlog>=24.0.0
langsmith>=0.1.0       # optional — only needed for LangSmith tracing

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-xdist>=3.5.0    # for -n parallel runs
httpx>=0.27.0           # for FastAPI AsyncClient in tests
